+++ ThirdPartyAdapters/line/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next Version
 - Implemented AdLoader to enable RTB for Interstitial Ads.
+- Implemented AdLoader to enable RTB for Rewarded Ads.
 
 #### Version 2.8.20240827.0
 - Verified compatibility with FiveAd SDK version 2.8.20240827.0